@@ -1,5 +1,5 @@
 import { Injectable } from '@angular/core';
-import { CanDeactivate } from '@angular/router';
+
 import { Observable } from 'rxjs';
 
 export interface ComponentCanDeactivate {
@@ -10,7 +10,7 @@ export interface ComponentCanDeactivate {
 @Injectable({
   providedIn: 'root'
 })
-export class PendingChangesGuard implements CanDeactivate<ComponentCanDeactivate> {
+export class PendingChangesGuard  {
   canDeactivate(component: ComponentCanDeactivate): boolean | Observable<boolean> {
     // If the form is not dirty, exit without modal.
     if (component.canDeactivate()) {
