@@ -2,9 +2,10 @@ import { Component, OnInit } from '@angular/core';
 import { Location } from '@angular/common';
 
 @Component({
-  selector: 'forbidden',
-  templateUrl: './forbidden.component.html',
-  styleUrls: ['./forbidden.component.scss']
+    selector: 'forbidden',
+    templateUrl: './forbidden.component.html',
+    styleUrls: ['./forbidden.component.scss'],
+    standalone: false
 })
 export class ForbiddenComponent implements OnInit {
 
