@@ -1,9 +1,10 @@
 import { Component, OnInit } from '@angular/core';
 
 @Component({
-  selector: 'app-error',
-  templateUrl: './page-404.component.html',
-  styleUrls: ['./page-404.component.scss']
+    selector: 'app-error',
+    templateUrl: './page-404.component.html',
+    styleUrls: ['./page-404.component.scss'],
+    standalone: false
 })
 export class Page404Component implements OnInit {
 
