@@ -4,9 +4,10 @@ import { AuthappService } from 'src/app/core/services/authapp.service';
 import { UserStateService } from 'src/app/core/services/user-state.service';
 
 @Component({
-  selector: 'header',
-  templateUrl: './header.component.html',
-  styleUrls: ['./header.component.scss']
+    selector: 'header',
+    templateUrl: './header.component.html',
+    styleUrls: ['./header.component.scss'],
+    standalone: false
 })
 export class HeaderComponent implements OnInit {
 
