@@ -1,9 +1,10 @@
 import { Component, Input, OnInit } from '@angular/core';
 
 @Component({
-  selector: 'app-jumbotron',
-  templateUrl: './jumbotron.component.html',
-  styleUrls: ['./jumbotron.component.scss']
+    selector: 'app-jumbotron',
+    templateUrl: './jumbotron.component.html',
+    styleUrls: ['./jumbotron.component.scss'],
+    standalone: false
 })
 export class JumbotronComponent implements OnInit {
 
