@@ -4,17 +4,17 @@ import { Ingredients } from "./Ingredients";
 import { Vat } from "./Vat";
 
 export interface ArticleResponse {
-  codart: string,
-  description: string,
-  um: string,
-  codStat: string,
-  pcsCart: number,
-  netWeight: number,
-  idArtStatus: number,
-  price: number,
-  creationDate: Date,
-  category: Category,
-  vat: Vat,
-  barcodes: Barcode[],
-  ingredients: Ingredients
+  codart?: string,
+  description?: string,
+  um?: string,
+  codStat?: string,
+  pcsCart?: number,
+  netWeight?: number,
+  idArtStatus?: number,
+  price?: number,
+  creationDate?: string,
+  category?: Category,
+  vat?: Vat,
+  barcodes: Barcode[], // Always exists (if empty => [])
+  ingredients?: Ingredients
 }
