@@ -2,11 +2,12 @@ import { Pipe, PipeTransform } from '@angular/core';
 import { ItemStatus } from '../enums';
 
 @Pipe({
-  name: 'literalItemStatus'
+    name: 'literalItemStatus',
+    standalone: false
 })
 export class LiteralItemStatusPipe implements PipeTransform {
 
-  transform(numItemStatus: number | null): unknown {
+  transform(numItemStatus: number | null | undefined): unknown {
     const mapArtStatus: Record<number, ItemStatus> = {
       1: ItemStatus.Active,
       2: ItemStatus.Suspended,
